test: add service type extension tests
