+++ DesktopApplicationTemplate.Tests/CreateServiceViewModelTests.cs
@@ -1,6 +1,5 @@
 using DesktopApplicationTemplate.UI.ViewModels;
 using DesktopApplicationTemplate.Models;
-using DesktopApplicationTemplate.Core.Converters;
 using Xunit;
 
 namespace DesktopApplicationTemplate.Tests
@@ -16,10 +15,10 @@ namespace DesktopApplicationTemplate.Tests
         [InlineData(ServiceType.Ftp)]
         public void GenerateDefaultName_ReturnsIncrementedName(ServiceType type)
         {
-            var existing = new[] { $"{ServiceTypeJsonConverter.ToLegacyString(type)}1" };
+            var existing = new[] { $"{type.ToLegacyString()}1" };
             var vm = new CreateServiceViewModel(existing);
             var name = vm.GenerateDefaultName(type);
-            Assert.Equal($"{ServiceTypeJsonConverter.ToLegacyString(type)}2", name);
+            Assert.Equal($"{type.ToLegacyString()}2", name);
             ConsoleTestLogger.LogPass();
         }
     }
