+++ DesktopApplicationTemplate.Tests/TestHelpers.cs
@@ -1,4 +1,3 @@
-using DesktopApplicationTemplate.Core.Converters;
 using DesktopApplicationTemplate.Models;
 using DesktopApplicationTemplate.UI.ViewModels;
 
@@ -9,6 +8,6 @@ public static class TestHelpers
     public static ServiceListModel CreateService(ServiceType type, string name) => new ServiceListModel
     {
         ServiceType = type,
-        DisplayName = $"{ServiceTypeJsonConverter.ToLegacyString(type)} - {name}"
+        DisplayName = $"{type.ToLegacyString()} - {name}"
     };
 }
